+++ lib/src/editor/raw_editor/raw_editor_state.dart
@@ -1201,8 +1201,6 @@ class QuillRawEditorState extends EditorState
       } else {
         _showCaretOnScreen();
       }
-    } else {
-      widget.config.focusNode.requestFocus();
     }
   }