+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `copyWith` methods to `HorizontalSpacing`, `VerticalSpacing`, `DefaultTextBlockStyle`, and `DefaultListBlockStyle` for immutable updates of properties [#2550](https://github.com/singerdmx/flutter-quill/pull/2550).
 - Finnish (fi) language translation [#2564](https://github.com/singerdmx/flutter-quill/pull/2564).
 
+=======
+### Fixed
+
+ - Move focus request out of requestKeyboard to avoid focus ping pong between editors.
+
 ## [11.4.0] - 2025-04-23
 
 ### Added